change parameter type fromm uint8 to uint32 since syscfg requires 32 bit aligned words

+++ syscfg.h
@@ -27,6 +27,6 @@ struct syscfg {
 #define EXTI_PORT_H 7UL
 #define EXTI_PORT_I 8UL
 
-void cfg_exti_interrupt(uint8_t port, uint8_t pin);
+void cfg_exti_interrupt(uint32_t port, uint32_t pin);
 
 #endif
